finish pipe and redirection
