+++ Makefile
@@ -10,7 +10,7 @@ INCLUDE_DIR = ./include
 MAIN = main signal
 CMD = cd echo env exec exit export pwd unset
 PARSER =  arg_splitter asterisk_splitter expand_envvar expand_wildcard get_file_list parser rm_quotes
-REDIR = cmd_redir exec_redir utils_redir
+REDIR = cmd_redir exec_redir utils_redir utils_redir2
 PIPE = utils_pipe utils
 
 SRCS =	$(addsuffix .c, $(MAIN)) \
