+++ redirection/utils_redir2.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils_redir.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kyung-ki <kyung-ki@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 1970/01/01 00:00:00 by kyung-ki          #+#    #+#             */
+/*   Updated: 2024/01/23 19:49:07 by kyung-ki         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../include/minishell.h"
+
+void	redir_excute(char **args, char **envp, t_node *node)
+{
+	if (node->redir_flag)
+	{
+		if (exec_redir(args, envp, node))
+			return ;
+	}
+}
